start function implementation
This is going to take some thought.
How do we `call` a function? goto?
This implies goto (the language construct) is different from goto (used
by the VM implementation) because the language's goto does not allow
jumping out of the current function.

I suppose `call` could get a privileged version of goto that ignores
this rule...